# docs/api/openapi.yml
...
openapi: 3.1.0
info:
  title: Elevation of Privilege — threat modelling card game
  description: |
    HTTP API for the digital Elevation of Privilege threat modelling card game.

    The card deck is reference data: it is seeded by a database migration and is
    never mutated through this API. There is deliberately no POST, PUT, PATCH or
    DELETE on `/api/v1/cards`.

    The deck is the real Elevation of Privilege deck: 78 cards, thirteen ranks
    in each of the six STRIDE suits. © 2009 Microsoft Corporation, licensed
    Creative Commons Attribution 3.0 United States — see
    `docs/EoP_Microsoft_Docs/LICENCE-eop-deck.md`. Attribution is the only
    obligation the licence imposes, and it is discharged in the running
    application rather than only in this repository.

    ## Sessions and state authority

    A game is played in a session. A facilitator creates one and shares its join
    code; players join with that code and a display name. Every participant then
    holds an opaque identity token (ADR-015) sent in the `X-EoP-Player-Token`
    header. There is no account system and no authentication scheme, so that
    token is the only thing distinguishing one player from another.

    **The database is the only authority on state.** `GET /api/v1/sessions/{sessionId}`
    is the single endpoint that reports it, and the event stream at
    `/api/v1/sessions/{sessionId}/events` only announces *that* something changed
...

...
  version: 1.0.0-SNAPSHOT
  license:
...
servers:
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [sessions]
      summary: Create a game session
      description: |
        Creates a session in `LOBBY` status and admits the caller as its
        `FACILITATOR` at `seatOrder` 0.

        The response is the only time the identity token is ever transmitted in
        plaintext. A client that loses it has lost that player identity and must
        join again as a new player.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: The session was created and the caller admitted as facilitator.
          headers:
            Location:
              description: URI of the created session.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionAdmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/sessions/{sessionId}/leaderboard:
    get:
      tags: [game-over]
      summary: Read the final leaderboard
      description: |
        The final leaderboard for a completed session: one row per seated player,
        sorted by total score descending, with a per-STRIDE breakdown of captured
        cards.

        This endpoint is gated by `eop.features.game-over` (ADR-013). While that
        flag is off the handler does not exist and the path returns 404.

        Any seated player may read it. The leaderboard names only cards that have
        already been played face up, so it discloses nothing the table cannot
        already see. A stranger who guesses a session identifier is refused with
        `403` because they hold no credential for that session.

        The leaderboard is available in any session state, not only `COMPLETED`.
        Before the game ends it reflects the score so far; after it ends it is
        the final result.

        Players with equal scores share the same rank (competition ranking). The
        `tied` flag makes this visible to a client rendering an ordinal without
        having to compare totals itself.
      operationId: getLeaderboard
      security:
        - playerToken: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Identifier of the session.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The leaderboard of the session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '500':
          description: |
            The stored game contradicts itself and no leaderboard can be derived
            from it. The body deliberately names nothing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/sessions/{sessionId}/new-game:
    post:
      tags: [game-over]
      summary: Start a new game with the same players
      description: |
        Clears all tricks and hands, re-deals the 68-card deck, and transitions
        the session back to `IN_PROGRESS`. All players remain in their original
        seats with the same seat numbers. The opening leader is the seat holding
        the 3 of Tampering.

        This endpoint is gated by `eop.features.game-over` (ADR-013). While that
        flag is off the handler does not exist and the path returns 404.

        Only the facilitator may call this. Any other seated player receives
        `403`. A stranger receives `403` for the same reason as every other
        operation.

        The session must be in `COMPLETED` status. Calling this while the session
        is `IN_PROGRESS` or `LOBBY` returns `409`.

        On success a `hand-dealt` event is published to the session's event
        stream, so all connected clients learn that the new hand is ready without
        polling.
      operationId: startNewGame
      security:
        - playerToken: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Identifier of the session.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: |
            The new game was started. Clients should re-read the session state
            and their own hand.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '409':
          description: |
            The session is not in a state where a new game can be started. The
            `detail` field names the current status.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  schemas:
...
    ProblemDetail:
      type: object
      description: |
        RFC 9457 problem details (ADR-005). Every 4xx and 5xx response uses this
        shape. Served as `application/problem+json`.

        The production profile suppresses exception classes, messages,
        stacktraces and binding errors, so `detail` is deliberately terse there.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
          default: about:blank
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code, repeated for clients that only read the body.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying this specific occurrence.
    SessionStatus:
      type: string
      description: |
        Lifecycle state of a session. Only `LOBBY` accepts new players, and only
        `LOBBY` can be started.
      enum:
        - LOBBY
        - IN_PROGRESS
        - COMPLETED
        - ABANDONED
    PlayerRole:
      type: string
      description: |
        `FACILITATOR` is whoever created the session. The role carries exactly one
        privilege — starting play — and is not transferable.
      enum:
        - FACILITATOR
        - PARTICIPANT
    ConnectionStatus:
      type: string
      description: |
        Whether the server currently believes this player is connected.

        Advisory only. A dead connection is discovered on the next failed write,
        so this can over-report `CONNECTED` between heartbeats. It is a display
        hint, never an input to a game rule.
      enum:
        - CONNECTED
        - DISCONNECTED
    DisplayName:
      type: string
      minLength: 1
      maxLength: 40
      description: |
        The name a player is known by at the table. Free text, unverified, and
        not unique — nothing stops two players choosing the same name. It is
        length-bounded and rejected if blank or containing control characters,
        because "anonymous" does not mean "unvalidated" (ADR-015).
      examples: ['Ada']
    JoinCode:
      type: string
      minLength: 6
      maxLength: 6
      pattern: '^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{6}$'
      description: |
        Six characters of Crockford base32, drawn from a cryptographically secure
        generator. The alphabet excludes `I`, `L`, `O` and `U`, so a code cannot
        be misread as a different valid code and cannot spell an unfortunate
        word.

        Always emitted in this canonical upper-case form. Input is normalised
        before lookup, so a code typed in lower case or with `I`/`L`/`O`
        substituted still resolves.
      examples: ['7QK2FM']
    CreateSessionRequest:
      type: object
      description: Everything needed to open a session — a name, and nothing else.
      required: [displayName]
      properties:
        displayName:
          $ref: '#/components/schemas/DisplayName'
...
    Player:
      type: object
      description: |
        One player at the table.

        Note what is absent: the stored hash of the player's identity token. It
        is never serialised, to any caller, under any flag. The only plaintext
        token any client ever sees is its own, in the admission response.
      required: [playerId, displayName, seatOrder, role, connectionStatus]
      properties:
        playerId:
          type: string
          format: uuid
          description: Stable identifier, assigned at join.
        displayName:
          $ref: '#/components/schemas/DisplayName'
        seatOrder:
          type: integer
          format: int32
          minimum: 0
          maximum: 5
          description: |
            Position at the table, assigned once at join and never recomputed.
            Play is clockwise, so "who plays next" is derived from this: it is
            game data, not presentation ordering. It is not re-derived from a
            database sort or a join timestamp, because a reconnect must not move
            a player around the table.
        role:
          $ref: '#/components/schemas/PlayerRole'
        connectionStatus:
          $ref: '#/components/schemas/ConnectionStatus'
    SessionState:
      type: object
      description: |
        The complete state of a session as stored in the database. This is the
        only representation of state the API produces, and the only thing a
        reconnecting client needs to read.
      required: [sessionId, joinCode, status, players, createdAt, updatedAt]
      properties:
        sessionId:
          type: string
          format: uuid
          description: |
            Identifier of the session. Appears in shareable URLs and is not a
            secret — holding it grants nothing without a token.
        joinCode:
          $ref: '#/components/schemas/JoinCode'
        status:
          $ref: '#/components/schemas/SessionStatus'
        players:
          type: array
          minItems: 1
          maxItems: 6
          description: |
            Everyone admitted so far, in ascending `seatOrder`. Three is the
            minimum to start and six is the maximum the deck and the game
            support.
          items:
            $ref: '#/components/schemas/Player'
        createdAt:
          type: string
          format: date-time
          description: When the session was created.
        updatedAt:
          type: string
          format: date-time
          description: When the session last changed.
    SessionAdmission:
      type: object
      description: |
        Returned by both create and join. One schema rather than two, because the
        two payloads are identical and two names for one shape invite drift.

        This is the only response that carries a plaintext identity token.
      required: [playerToken, playerId, session]
      properties:
        playerToken:
          type: string
          description: |
            The caller's opaque identity credential, base64url-encoded (ADR-015).
            Sent back in `X-EoP-Player-Token` on every subsequent request. The
            server stores only a SHA-256 hash of it and cannot reissue it, so a
            client that loses it has lost that player identity.
          examples: ['t7Yc0Qw9Jt5mB3nR8vK1sZ2xL6hD4fA0pE9gU7iO3bY']
        playerId:
          type: string
          format: uuid
          description: |
            Which player in `session.players` the caller is. Supplied so the
            client does not have to infer its own identity by matching on display
            name, which is not unique.
        session:
          $ref: '#/components/schemas/SessionState'
...
    LeaderboardRow:
      type: object
      description: |
        One player's row in the final leaderboard: their rank, name, total score,
        and a breakdown of captured cards by STRIDE category.

        `capturedBySuit` counts the cards the player captured (i.e. won the trick
        with) in each STRIDE category. A card the player played but did not win
        the trick with does not appear here. The map always contains all six
        STRIDE categories, with zero for any category where the player captured
        nothing.
      required: [playerId, seatOrder, displayName, points, position, tied, capturedBySuit]
      properties:
        playerId:
          type: string
          format: uuid
          description: The player this row belongs to.
        seatOrder:
          type: integer
          format: int32
          minimum: 0
          maximum: 5
          description: The seat they hold.
        displayName:
          allOf:
            - $ref: '#/components/schemas/DisplayName'
          description: The name they chose, for display only.
        points:
          type: integer
          format: int32
          minimum: 0
          description: Their total score across the whole game.
        position:
          type: integer
          format: int32
          minimum: 1
          description: |
            Their place, competition ranked. Players level on points share a
            position, and the position after a shared one skips accordingly.
        tied:
          type: boolean
          description: Whether another player holds the same total.
        capturedBySuit:
          type: object
          description: |
            Count of cards captured (trick wins) per STRIDE category. All six
            categories are always present; zero means none captured.
          required:
            - SPOOFING
            - TAMPERING
            - REPUDIATION
            - INFORMATION_DISCLOSURE
            - DENIAL_OF_SERVICE
            - ELEVATION_OF_PRIVILEGE
          properties:
            SPOOFING:
              type: integer
              format: int32
              minimum: 0
            TAMPERING:
              type: integer
              format: int32
              minimum: 0
            REPUDIATION:
              type: integer
              format: int32
              minimum: 0
            INFORMATION_DISCLOSURE:
              type: integer
              format: int32
              minimum: 0
            DENIAL_OF_SERVICE:
              type: integer
              format: int32
              minimum: 0
            ELEVATION_OF_PRIVILEGE:
              type: integer
              format: int32
              minimum: 0
    Leaderboard:
      type: object
      description: |
        The leaderboard for a session: one row per seated player, sorted by total
        score descending (highest first). Players with equal scores share the same
        rank.

        This is the Game Over screen's primary data. It is available in any
        session state, not only `COMPLETED`.
      required: [rows, sessionStatus]
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 6
          items:
            $ref: '#/components/schemas/LeaderboardRow'
          description: |
            One row per seated player, ordered by position and, among players
            level on points, by seat.
        sessionStatus:
          $ref: '#/components/schemas/SessionStatus'
...
  responses:
    BadRequest:
      description: |
        The request was malformed — for example a `cardId` that is not a UUID, a
        body that does not parse, or a display name that is blank or longer than
        forty characters.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    Forbidden:
      description: |
        The `X-EoP-Player-Token` header is missing, unrecognised, or belongs to a
        different session — or it is valid but the player it identifies is not
        permitted to perform this operation — or the session has expired.

        403 rather than 401, deliberately. There is no authentication scheme
        here: no realm, no challenge, nothing a client could retry differently
        after being told to authenticate. A 401 is required to carry a
        `WWW-Authenticate` header, and emitting one would advertise a challenge
        that does not exist. The request was understood and refused, which is
        what 403 means. Session expiry is also a 403: the credential was once
        valid but the session it belonged to no longer exists. The `title` field
        distinguishes the cases: `"Player not recognised"` for an invalid token,
        `"Session expired"` for an expired session (ADR-036).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    SessionNotFound:
      description: |
        No session exists with that identifier.

        Note the asymmetry with joining by code: a `sessionId` is a UUID that
        nobody can usefully guess, so there is nothing to be gained by hiding
        whether one exists. A join code is six characters, so that endpoint
        deliberately does hide it.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    TooManyRequests:
      description: |
        Too many requests from this source address and further attempts are
        refused for a period.

        Three controls can produce this response:

        - **Session creation limiter** — a single address may create at most
          `eop.web.session-creation-limit` sessions (default 5) per 60-second
          window. The slot is reserved before any database work, so a refused
          request never commits a row (ADR-033).
        - **Join attempt limiter** — too many failed join attempts from this
          source or against this code. A six-character code has roughly thirty
          bits of entropy, which is only unguessable while guessing is
          rate-limited — so this limiter is a primary security control rather
          than a courtesy (ADR-019).
        - **Subscriber cap** — too many concurrent SSE subscribers for a session
          or globally. The cap is not time-bounded, so no `Retry-After` header
          is emitted for this case (EOP-20, ADR-034).
      headers:
        Retry-After:
          description: |
            Seconds to wait before attempting again.
            Absent for subscriber-cap refusals, which are not time-bounded.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    RequestEntityTooLarge:
      description: |
        The request body exceeded the 16 KB limit enforced by the reverse proxy
        (Caddy `request_body { max_size 16KB }` on `/api/*`).

        This response is produced by the proxy, not the application. The body
        is a plain-text or empty response, not a problem detail.
    ServiceUnavailable:
      description: |
        No new lobby could be opened, because every join code the generator
        offered was already taken by a live session. The request was valid and
        nothing malfunctioned, so this is a capacity statement rather than an
        error: the same request is expected to succeed shortly.

        A six-character code carries roughly thirty bits of entropy, so a run of
        collisions across the whole attempt budget means the number of live
        lobbies is approaching what that code space comfortably bears. Unlike
        429, a caller cannot provoke this — which is why each occurrence is
        logged as an operational signal (EOP-17).

        The `detail` says nothing about join codes. The collision is between our
        own generator and our own data, and a caller who never sees a code being
        generated has no use for that mechanism.
      headers:
        Retry-After:
          description: Seconds to wait before attempting again.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  securitySchemes:
    playerToken:
      type: apiKey
      in: header
      name: X-EoP-Player-Token
      description: |
        The opaque identity credential issued at create or join (ADR-015).

        A custom header rather than `Authorization: Bearer`, because this is not
        OAuth and naming it so would invite the assumption that it behaves like
        one.

        It is **never** accepted as a query parameter, including on the event
        stream where `EventSource` would have made that convenient. A credential
        in a query string is a credential in the access log, in the browser
        history and on screen during a shared call. Clients that need to stream
        use `fetch`-based streaming, which can set headers.